backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.34
requests==2.32.3
python-dotenv==1.0.1

# Auth (JWT + password hashing)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0

# Multipart form-data support (required for Form/File uploads)
python-multipart==0.0.9

# Resume PDF extraction
pypdf==4.3.1
PyPDF2==3.0.1
pdfminer.six==20240706

# OCR fallback for scanned PDFs
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.4.0
pymysql==1.1.1